requirements.txt
python-dotenv
openai
langchain-openai
langchain-community
multion